[pf-372] Delete legacy version of assessment controller get tests
